+++ CHANGELOG.md
@@ -6,4 +6,5 @@ v0.7.0
 ----
  * tabbar: New interface with index.
  * page: Fixed bug with modifier not being applied to pages. Add contentStyle property.
+ * page: Add `renderFixed` prop to render fixed position conent.
  * range: Fix `value` prop.